refactor: 💡 completely changing the build system and site setup
Original resume site is still a static html site at the moment (early
Dec 2025). A later attempt to setup a gatsby + netlify version of the
site was never completed to a level worth deploying.  This change
replaces the netlify commits with the beginnings of a hugo site which I
intend to deploy on vercel.

BREAKING CHANGE: 🧨 changed from gatsby to hugo